Integrate V4 session library and namespaces.

+++ upload/system/library/session/file.php
@@ -1,36 +1,46 @@
 <?php
 namespace Session;
 class File {
-	private $directory;
-
+	public function __construct(Registry $registry) {
+		$this->config = $registry->get('config');
+	}
+	
 	public function read($session_id) {
-		$file = DIR_SESSION . '/sess_' . basename($session_id);
+		$file = DIR_SESSION . 'sess_' . basename($session_id);
 
 		if (is_file($file)) {
-			$handle = fopen($file, 'r');
+			$size = filesize($file);
 
-			flock($handle, LOCK_SH);
+			if ($size) {
+				$handle = fopen($file, 'r');
 
-			$data = fread($handle, filesize($file));
+				flock($handle, LOCK_SH);
 
-			flock($handle, LOCK_UN);
+				$data = fread($handle, $size);
 
-			fclose($handle);
+				flock($handle, LOCK_UN);
 
-			return unserialize($data);
-		} else {
-			return array();
+				fclose($handle);
+
+				return json_decode($data, true);
+			} else {
+				return array();
+			}
 		}
+
+		return array();
 	}
 
 	public function write($session_id, $data) {
-		$file = DIR_SESSION . '/sess_' . basename($session_id);
+		$file = DIR_SESSION . 'sess_' . basename($session_id);
 
-		$handle = fopen($file, 'w');
+		$handle = fopen($file, 'c');
 
 		flock($handle, LOCK_EX);
 
-		fwrite($handle, serialize($data));
+		fwrite($handle, json_encode($data));
+
+		ftruncate($handle, ftell($handle));
 
 		fflush($handle);
 
@@ -42,33 +52,21 @@ class File {
 	}
 
 	public function destroy($session_id) {
-		$file = DIR_SESSION . '/sess_' . basename($session_id);
+		$file = DIR_SESSION . 'sess_' . basename($session_id);
 
 		if (is_file($file)) {
 			unlink($file);
 		}
 	}
 
-	public function __destruct() {
-		if (ini_get('session.gc_divisor')) {
-			$gc_divisor = ini_get('session.gc_divisor');
-		} else {
-			$gc_divisor = 1;
-		}
+	public function gc() {
+		if (round(rand(1, $this->config->get('session_divisor') / $this->config->get('session_probability'))) == 1) {
+			$expire = time() - $this->config->get('session_expire');
 
-		if (ini_get('session.gc_probability')) {
-			$gc_probability = ini_get('session.gc_probability');
-		} else {
-			$gc_probability = 1;
-		}
-
-		if ((rand() % $gc_divisor) < $gc_probability) {
-			$expire = time() - ini_get('session.gc_maxlifetime');
-
-			$files = glob(DIR_SESSION . '/sess_*');
+			$files = glob(DIR_SESSION . 'sess_*');
 
 			foreach ($files as $file) {
-				if (filemtime($file) < $expire) {
+				if (is_file($file) && filemtime($file) < $expire) {
 					unlink($file);
 				}
 			}
